requirements.txt
async-timeout==4.0.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
numpy==1.24.3
opencv-python==4.7.0.72
passlib==1.7.4
protobuf==3.20.3
PyJWT==2.6.0
python-dotenv==1.0.0
redis==4.5.5
six==1.16.0
Werkzeug==2.3.2
zipp==3.15.0
